*** empty log message ***
svn path=/trunk/boinc/; revision=4567

+++ lib/gui_rpc_client.h
@@ -60,7 +60,7 @@ public:
     int nrpc_failures;          // # of consecutive times we've failed to
                                 // contact all scheduling servers
     int master_fetch_failures;
-    int min_rpc_time;           // earliest time to contact any server
+    double min_rpc_time;           // earliest time to contact any server
 
     bool master_url_fetch_pending; // need to fetch and parse the master URL
     bool sched_rpc_pending;     // contact scheduling server for preferences
